Moved checks
